chore(connect-explorer-theme): remove circular imports
