+++ packages/connect-explorer-theme/src/contexts/sidebar-focus.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+
+export const FocusedItemContext = createContext<null | string>(null);
+export const OnFocusItemContext = createContext<null | ((item: string | null) => any)>(null);
